Refactor CurrentSiteContext to utilize utility functions for website YAML handling and improve code clarity
