+++ src/client/test/Integration/power-pages/actions-hub/ActionsHubTreeDataProvider.test.ts
@@ -123,9 +123,6 @@ describe("ActionsHubTreeDataProvider", () => {
             ] as IWebsiteDetails[];
             sinon.stub(CommandHandlers, 'fetchWebsites').resolves({ activeSites: mockActiveSites, inactiveSites: mockInactiveSites });
 
-            sinon.stub(WebsiteUtil, 'getActiveWebsites').resolves(mockActiveSites);
-            sinon.stub(WebsiteUtil, 'getAllWebsites').resolves(mockAllSites);
-
             PacContext['_authInfo'] = null;
             const provider = ActionsHubTreeDataProvider.initialize(context, pacTerminal);
             const result = await provider.getChildren();
@@ -191,11 +188,17 @@ describe("ActionsHubTreeDataProvider", () => {
             const environmentGroup = result![0] as EnvironmentGroupTreeItem;
             expect(environmentGroup.environmentInfo.currentEnvironmentName).to.equal("TestOrg");
             expect(environmentGroup['_activeSites']).to.deep.equal(mockActiveSites);
-            expect(environmentGroup['_inactiveSites']).to.deep.equal([mockAllSites[1]]);
+            expect(environmentGroup['_inactiveSites']).to.deep.equal(mockInactiveSites);
         });
 
         it("should return empty array when auth info is not available", async () => {
             sinon.stub(PacContext, "AuthInfo").get(() => null);
+
+            const provider = ActionsHubTreeDataProvider.initialize(context, pacTerminal);
+            const result = await provider.getChildren();
+
+            expect(result).to.be.an('array').that.is.empty;
+        });
 
         it("should return environment group tree item with default name when no auth info is available", async () => {
             const mockActiveSites = [
